Merging v0.1.4a_dev into the main branch.

+++ prism/tests/integration/integration_test_class.py
@@ -52,8 +52,12 @@ class IntegrationTestCase(unittest.TestCase):
 
 
     def _set_up_wkdir(self):
+        # Remove logs.log from project
+        if Path(Path.cwd() / 'logs.log').is_file():
+            os.unlink(Path.cwd() / 'logs.log')
+        
         os.chdir(TEST_PROJECTS)
-
+        
 
     def _is_valid_project(self, path):
         """
@@ -175,7 +179,7 @@ class IntegrationTestCase(unittest.TestCase):
         Remove file outputs from `output` folder of project
         """
         for file in Path(wkdir / 'output').iterdir():
-            if Path(wkdir / 'output' / file).is_file():
+            if Path(wkdir / 'output' / file).is_file() and file.name!=".exists":
                 os.unlink(file)
     
 
@@ -231,6 +235,24 @@ class IntegrationTestCase(unittest.TestCase):
         if_name_main_block = module_ast_tree.body[-1]
         self.assertTrue(isinstance(if_name_main_block, ast.If))
         return astor.to_source(if_name_main_block)
+
+    
+    def _remove_profile_yml(self, wkdir):
+        """
+        Remove the profile.yml file, if it exists
+        """
+        if Path(wkdir / 'profile.yml').is_file():
+            os.unlink(Path(wkdir / 'profile.yml'))
+    
+
+    def _profile_yml_as_dict(self, wkdir):
+        """
+        Open the profile.yml file as a dict
+        """
+        with open(Path(wkdir / 'profile.yml'), 'r') as f:
+            yml_dict = yaml.safe_load(f)
+        f.close()
+        return yml_dict
 
 
 # EOF
